+++ paddle/fluid/imperative/type_defs.h
@@ -15,9 +15,8 @@ limitations under the License. */
 #pragma once
 
 #include <map>
+#include <memory>
 #include <string>
-#include <unordered_map>
-#include <utility>
 #include <vector>
 
 namespace paddle {
@@ -25,15 +24,10 @@ namespace imperative {
 
 class VarBase;
 class OpBase;
+class Tracer;
 
-typedef std::map<std::string, std::vector<VarBase*>> VarBasePtrMap;
-typedef std::map<std::string, std::vector<const VarBase*>> ConstVarBasePtrMap;
-typedef std::map<std::string, std::vector<OpBase*>> OpBasePtrMap;
-typedef std::unordered_map<
-    const VarBase*,
-    std::pair<platform::Place, std::vector<std::pair<int, VarBase*>>>>
-    BackwardSumMap;  // var_grad -> {place, {id -> var_grad@rename}}
-typedef std::unordered_map<const VarBase*, int> GradientRef;
+using NameVarBaseMap =
+    std::map<std::string, std::vector<std::shared_ptr<VarBase>>>;
 
 }  // namespace imperative
 }  // namespace paddle
